requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
audioread==3.0.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.2
GoogleAudio==0.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
librosa==0.9.1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
openai==1.35.7
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
requests==2.32.3
resampy==0.4.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
sympy==1.12.1
tbb==2021.13.0
threadpoolctl==3.5.0
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
yandex-speechkit==1.5.0
youtube-dl==2021.12.17
